+ Blogs & articles dynamic links with fetching data + <ArticleView />

+++ src/components/UserBlogsListInDashboard.js
@@ -1,51 +1,37 @@
 import { useState, useEffect } from "react";
-import { database } from "../firebase";
-import { ref, remove, onValue } from "firebase/database";
+import { useDatabase } from "../hooks/use-database";
+import { useAuth } from "../hooks/use-auth";
 import { Link, Switch, Route, useRouteMatch } from "react-router-dom";
 import BlogEditionForm from "./BlogEditionForm";
 
-export default function UserBlogsListInDashboard({ userId, setBlogKeyForNewArticle, setBlogTitleForNewArticle }) {
+export default function UserBlogsListInDashboard({
+    setBlogKeyForNewArticle,
+    setBlogTitleForNewArticle,
+    setCurrentBlogKey,
+    setCurrentBlogLink
+    }) {
 
     let {path, url} = useRouteMatch();
 
-    const [userBlogsList, setUserBlogsList] = useState(null);
+    const { blogs, deleteBlog } = useDatabase();
+    const { user } = useAuth();
 
-    const [currentBlogKey, setCurrentBlogKey] = useState(null);
-    const [currentBlogLink, setCurrentBlogLink] = useState(null);
-
-    function fetchUserBlogsList(userId) {
-        const userBlogsRef = ref(database, 'users/' + userId + '/blogs');
-        onValue(userBlogsRef, (snapshot) => {
-            if (snapshot) {
-                const fetchedUserBlogsList = snapshot.val();        
-                setUserBlogsList(fetchedUserBlogsList ? Object.entries(fetchedUserBlogsList) : null);
-                console.log("User blogs list fetched to dashboard", fetchedUserBlogsList);
-            }
-        });
-    }
-
-    function deleteBlog(blogKey) {
-        // eslint-disable-next-line no-restricted-globals
-        const wantToDeleteBlog = confirm("Are you sure, you want to delete this blog & all articles from this blog forever? There's no turning back... Delete blog?");
-        if (wantToDeleteBlog) {
-            remove(ref(database, 'users/' + userId + '/blogs/' + blogKey)).then(() => {
-                console.log("blog was deleted");
-                fetchUserBlogsList(userId);
-            }).catch((error) => {
-                // An error ocurred
-                console.log(error.message);
-            });
-            remove(ref(database, 'blogs/' + blogKey));
-        }
-    }
-
-    function convertBlogTitleIntoLink(blogTitle) {
-        return ("/" + blogTitle.replace(/ /g, "-").toLowerCase());
-    }
+    const [userBlogs, setUserBlogs] = useState(null);
 
     useEffect(() => {
-        fetchUserBlogsList(userId);
-    }, [userId]);
+        if (blogs && user) {
+            console.log(blogs);
+            const fetchedBlogs = Object.entries(blogs);
+            const currentUserBLogs = fetchedBlogs.filter(blog => blog[1].userId === user.uid)
+            setUserBlogs(currentUserBLogs);
+        } else {
+            console.log("there are no user and blogs...")
+        }
+    }, [blogs, user]);
+
+    function convertBlogTitleIntoLink(blogTitle) {
+        return (blogTitle.replace(/ /g, "-").toLowerCase());
+    }
 
     return (
         <Switch>
@@ -57,8 +43,8 @@ export default function UserBlogsListInDashboard({ userId, setBlogKeyForNewArtic
                         className="btn btn-info d-block my-3"
                     >Create new blog</Link>
                     {
-                        userBlogsList && userBlogsList.length ?
-                            userBlogsList.map((blog) =>
+                        userBlogs && userBlogs.length ?
+                            userBlogs.map((blog) =>
                                 <div className="container" key={blog[0]}>
                                     <div className="row">
                                         <div className="col">
@@ -84,10 +70,24 @@ export default function UserBlogsListInDashboard({ userId, setBlogKeyForNewArtic
                                                 <i className="bi bi-plus-square me-2" />
                                             </Link>
                                             <Link
+                                                className="text-primary"
+                                                to={`/blogs/${convertBlogTitleIntoLink(blog[1].title)}`}
+                                                onClick={() => {
+                                                    setCurrentBlogKey(blog[0]);
+                                                    setCurrentBlogLink(`/blogs${convertBlogTitleIntoLink(blog[1].title)}`);
+                                                }}
+                                            >
+                                                <i className="bi bi-eye me-2" />
+                                            </Link>
+                                            <Link
                                                 className="text-danger"
                                                 to="/dashboard/user-blogs"
                                                 onClick={() => {
-                                                    deleteBlog(blog[0]);
+                                                    // eslint-disable-next-line no-restricted-globals
+                                                    const wantToDeleteBlog = confirm("Are you sure, you want to delete this blog & all articles from this blog forever? There's no turning back... Delete blog?");
+                                                    if (wantToDeleteBlog) {
+                                                        deleteBlog(blog[0]);
+                                                    }
                                                 }}
                                             >
                                                 <i className="bi bi-trash" />
@@ -106,13 +106,11 @@ export default function UserBlogsListInDashboard({ userId, setBlogKeyForNewArtic
                     }
                 </div>
             </Route>
-            <Route path={currentBlogLink}>
+            {/*<Route path={currentBlogLink}>
                 <BlogEditionForm
-                    userId={userId}
                     blogKey={currentBlogKey}
-                    deleteBlog={deleteBlog}
                 />
-            </Route>
+            </Route>*/}
         </Switch>
     );
 }
